Implement getNodeById
Wrote this function for Edit.jsx in the front end, because my initial
implementation that didn't require a call to the database upon opening
the page, started acting up when I tried to repair the touch events from
D3 and React interfering with each other. I also thought being able
to share edit pages might be beneficial for some users.